DDF-1831 Refactored RegistryObjectMeacardType out of api bundle into registry common jar
