+++ catalog/registry/catalog-registry-api-impl/src/main/java/ddf/catalog/registry/api/impl/RegistryStoreImpl.java
@@ -50,8 +50,8 @@ import ddf.catalog.operation.UpdateRequest;
 import ddf.catalog.operation.UpdateResponse;
 import ddf.catalog.operation.impl.SourceResponseImpl;
 import ddf.catalog.registry.api.RegistryStore;
-import ddf.catalog.registry.api.metacard.RegistryObjectMetacardType;
 import ddf.catalog.registry.common.RegistryConstants;
+import ddf.catalog.registry.common.metacard.RegistryObjectMetacardType;
 import ddf.catalog.source.IngestException;
 import ddf.catalog.source.UnsupportedQueryException;
 import oasis.names.tc.ebxml_regrep.xsd.rim._3.ExternalIdentifierType;
