+++ catalog/registry/catalog-registry-api/pom.xml
@@ -33,10 +33,6 @@
             <groupId>ddf.catalog.core</groupId>
             <artifactId>catalog-core-api</artifactId>
         </dependency>
-        <dependency>
-            <groupId>ddf.catalog.core</groupId>
-            <artifactId>catalog-core-api-impl</artifactId>
-        </dependency>
     </dependencies>
 
     <build>
@@ -49,12 +45,8 @@
                     <instructions>
                         <Bundle-SymbolicName>${project.artifactId}</Bundle-SymbolicName>
                         <Export-Package>
-                            ddf.catalog.registry.api,
-                            ddf.catalog.registry.api.metacard
+                            ddf.catalog.registry.api
                         </Export-Package>
-                        <Embed-Dependency>
-                            catalog-core-api-impl
-                        </Embed-Dependency>
                         <Import-Package>
                             !org.codice.ddf.platform.util,
                             *
