+++ catalog/registry/catalog-registry-cswinputtransformer/src/test/java/ddf/catalog/registry/transformer/RegistryTransformerTest.java
@@ -44,7 +44,7 @@ import org.mockito.Mockito;
 import ddf.catalog.data.Metacard;
 import ddf.catalog.data.impl.AttributeImpl;
 import ddf.catalog.data.impl.MetacardImpl;
-import ddf.catalog.registry.api.metacard.RegistryObjectMetacardType;
+import ddf.catalog.registry.common.metacard.RegistryObjectMetacardType;
 import ddf.catalog.transform.CatalogTransformerException;
 import oasis.names.tc.ebxml_regrep.xsd.rim._3.RegistryObjectType;
